multi-level subcheck
