+++ AngularApp/projects/app-service-diagnostics/src/app/shared/components/step-views/check-step-view/check-step.component.html
@@ -1,19 +1,3 @@
 <div class="check-step step">
-    <div class="status-icon">
-        <status-icon [loading]="1" [status]="checkStepView.level | convertLevelToHealthStatus"></status-icon>
-    </div>
-    <div class="check-title bold">
-      {{checkStepView.title}}
-    </div>
-    <div class="details-link" *ngIf="checkStepView.subChecks.length>0"> <a tabindex="0" (click)="toggleSubChecks()"> <i class="{{expanded ? 'ib-dchevup' : 'ib-dchevdown'}}"></i> </a></div>
-    <div class="subchecks-container" *ngIf="expanded">
-      <div class="sub-check" *ngFor="let subCheck of checkStepView.subChecks; let i = index">
-        <div class="status-icon">
-          <status-icon [loading]="1" [status]="subCheck.level | convertLevelToHealthStatus"></status-icon>
-        </div>
-        <div class="check-title">
-          {{subCheck.title}}
-        </div>
-      </div>
-    </div>
+  <check [check]="checkStepView" [bold]="true"></check> 
 </div>
