+++ AngularApp/projects/app-service-diagnostics/src/app/shared/components/step-views/check-step-view/check.component.html
@@ -1,0 +1,14 @@
+<div class="check-container">
+    <div class="status-icon">
+        <status-icon [loading]="1" [status]="check.level | convertLevelToHealthStatus"></status-icon>
+    </div>
+    <div class="check-title {{bold ? 'bold':''}}">
+      {{check.title}}
+    </div>
+    <div class="details-link" *ngIf="check.subChecks.length>0"> <a tabindex="0" (click)="toggleSubChecks()"> <i class="{{expanded ? 'ib-dchevup' : 'ib-dchevdown'}}"></i> </a></div>
+    <div class="subchecks-container" *ngIf="expanded">
+      <div class="sub-check" *ngFor="let subCheck of check.subChecks; let i = index">
+        <check [check]="subCheck"></check> 
+      </div>
+    </div>
+</div>
