+++ AngularApp/projects/app-service-diagnostics/src/app/shared/components/step-views/check-step-view/check.component.scss
@@ -1,0 +1,45 @@
+.check-step{
+    padding: 9px 5px 9px 10px;
+
+    .check-title{
+        font-size: 13px;
+        margin-top: -20px;
+        margin-left: 30px;
+    }
+    
+    .bold{
+        font-weight: 500;
+    }
+
+    .subchecks-container{
+        padding-left: 25px;
+        .sub-check{
+            padding-top: 10px;
+        }
+    }
+
+    .details-link{
+        float: right;
+        margin-top: -16px;
+        padding-right: 5px;
+    }
+}
+
+.status-icon {
+    display: flex;
+    align-items: center;
+    padding-left: 1px;
+    width: 70px;
+    height: 100%;
+    .fa{
+        font-size: 20px !important;
+    }
+
+    .fa-exclamation-triangle{
+        margin-left: -2px;
+    }
+}
+
+
+
+
